requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
altair==5.5.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.1.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
azure-core==1.31.0
azure-functions==1.20.0
azure-identity==1.19.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.0
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.34.0
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
dirtyjson==1.0.8
distlib==0.3.9
distro==1.8.0
dnspython==2.7.0
docker==7.1.0
duckdb==0.8.1
dulwich==0.21.7
durationpy==0.9
elementpath==4.4.0
emoji==2.14.0
eval_type_backport==0.2.0
executing==2.0.1
Faker==19.13.0
fastapi==0.115.6
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphene==3.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.0.2
grpcio==1.68.1
h11==0.14.0
hiredis==3.0.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
ifcopenshell==0.7.10
ifctester==0.7.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
installer==0.7.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.12
langchain-chroma==0.1.4
langchain-community==0.3.12
langchain-core==0.3.25
langchain-huggingface==0.1.2
langchain-mongodb==0.3.0
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
langdetect==1.0.9
langsmith==0.2.3
lark==1.1.9
llama-index-core==0.11.17
llama-index-embeddings-adapter==0.2.2
llama-index-experimental==0.4.0
llama-index-finetuning==0.2.1
llama-index-llms-azure-openai==0.2.2
llama-index-llms-mistralai==0.2.6
llama-index-llms-openai==0.2.13
llama-index-postprocessor-cohere-rerank==0.2.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==1.1.0
mlflow==2.17.0
mlflow-skinny==2.17.0
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
more-itertools==10.5.0
motor==3.6.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.18.3
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==1.26.2
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.57.4
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.1
pandas==1.5.3
parameterized==0.9.0
parso==0.8.3
pathos==0.3.3
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
portalocker==2.10.1
posthog==3.7.4
pox==0.3.5
ppft==1.7.6.9
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.29.1
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.9.0
pymongo==4.9.2
pyparsing==3.1.1
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2023.3.post1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.11.4
sentence-transformers==3.2.0
shapely==2.0.5
shellingham==1.5.4
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==1.4.50
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomlkit==0.13.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.0
transformers==4.45.2
trove-classifiers==2024.10.16
typer==0.15.1
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.27.0
waitress==3.0.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.4
wrapt==1.16.0
xmlschema==3.3.2
yarl==1.15.2
zipp==3.20.2
